+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - SLI plugins support.
+- SLI `prometheus/v1` plugins.
 
 ## [v0.2.0] - 2021-05-24